requirements.txt
# AnomRadar v2 Dependencies
# Pinned versions for reproducibility

# CLI Framework
typer==0.9.0
click==8.1.7

# Configuration & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
tomli==2.0.1

# TUI Framework
textual==0.44.1

# HTTP & Network
httpx==0.25.1
dnspython==2.4.2
cryptography==41.0.7

# Export & Reporting
orjson==3.9.10
jinja2==3.1.2
markupsafe==2.1.3

# Utilities
rich==13.7.0
pyyaml==6.0.1

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
ruff==0.1.6
